Update to ATT-less APIC approach; disable use of the NodeID MSR for now

+++ node.h
@@ -30,7 +30,7 @@ public:
 	ht_t neigh_ht;
 	link_t neigh_link;
 	link_t neigh_sublink;
-	uint16_t apics[56];
+	uint32_t apics[56];
 	uint8_t napics;
 	unsigned nopterons;
 	unsigned cores;
